Binance integration first version
Former-commit-id: 7b5e7f921f7058a0a851d11b67c9668d33109224

+++ HFTPlatform/zHFT.InstructionBasedMarketClient.BitMex/BitMexInstructionBasedMarketClient.cs
@@ -325,7 +325,7 @@ namespace zHFT.InstructionBasedMarketClient.BitMex
                     WSMarketDataManager.SubscribeEvents(BaseManager._ORDERBOOK_L2, UpdateOrderBook);
                     WSMarketDataManager.SubscribeEvents(BaseManager._TRADE, UpdateTrade);
 
-                    ActiveSecurities = new Dictionary<int, Security>();
+                    ActiveSecurities = new Dictionary<long, Security>();
                     ContractsTimeStamps = new Dictionary<int, DateTime>();
                     return true;
                 }
